requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.4.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-image>=0.20.0
opencv-python>=4.7.0
tensorboard>=2.12.0
timm>=0.9.0

# For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0